requirements.txt
Flask==3.0.2
Pillow==10.4.0
rembg==2.0.58
openai==1.31.1